wydzielenie stora batteries
